+++ wear/src/main/java/com/asamm/locus/addon/wear/features/settings/MainSettingsActivity.kt
@@ -7,20 +7,41 @@ package com.asamm.locus.addon.wear.features.settings
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.View
-import android.widget.TextView
 import android.widget.Toast
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.wear.compose.material.Icon
+import androidx.wear.compose.material.MaterialTheme
+import androidx.wear.compose.material.ToggleChipDefaults
 import com.asamm.locus.addon.wear.R
 import com.asamm.locus.addon.wear.common.communication.DataPath
 import com.asamm.locus.addon.wear.common.communication.containers.DataPayload
 import com.asamm.locus.addon.wear.common.communication.containers.commands.EmptyCommand
 import com.asamm.locus.addon.wear.features.trackRecord.TrackRecordingService
-import com.asamm.locus.addon.wear.features.trackRecord.recording.sensors.RecordingSensorManager
+import com.asamm.locus.addon.wear.features.trackRecord.recording.sensors.BodySensorsManager
 import com.asamm.locus.addon.wear.gui.LocusWearActivity
+import com.asamm.locus.addon.wear.gui.compose.*
+import com.asamm.locus.addon.wear.gui.compose.theme.AppTheme
 import com.asamm.locus.addon.wear.utils.FeatureConfigEnum
-import com.google.android.material.switchmaterial.SwitchMaterial
+import locus.api.utils.Logger
 
-open class MainSettingsActivity : LocusWearActivity() {
+/**
+ * Main app settings activity.
+ */
+class MainSettingsActivity : LocusWearActivity() {
 
     // runs from the menu, not independently
     override val isChildLocusWearActivity: Boolean
@@ -35,102 +56,174 @@ open class MainSettingsActivity : LocusWearActivity() {
     override val isMakeHandshakeOnStart: Boolean
         get() = false
 
-    // switcher for HRM
-    private lateinit var switchHrm: SwitchMaterial
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main_settings)
-
-        // setup header
-        findViewById<TextView>(R.id.text_view_screen_header)
-                .text = getText(R.string.settings)
-
-        // set version text
-        val versionName = packageManager.getPackageInfo(packageName, 0).versionName
-        findViewById<TextView>(R.id.text_view_version)
-                .text = getString(R.string.version_name_X, versionName)
-
-        // set HRM toggle
-        switchHrm = findViewById(R.id.switch_hrm)
-        updateSwitchState()
+        setContent {
+            MainContent()
+        }
     }
 
-    private fun updateSwitchState() {
-        val hrmConfig = PreferencesEx.getHrmFeatureConfig()
-        switchHrm.isChecked = hrmConfig === FeatureConfigEnum.ENABLED
-        switchHrm.setTextColor(
-                if (hrmConfig !== FeatureConfigEnum.NOT_AVAILABLE) {
-                    getColor(R.color.base_dark_primary)
-                } else {
-                    getColor(R.color.base_light_disabled)
+    @Composable
+    fun MainContent() {
+        AppTheme {
+            Column {
+                Header()
+                Content()
+            }
+        }
+    }
+
+    @Composable
+    private fun Header() {
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(48.dp)
+                .background(MaterialTheme.colors.primary)
+                .padding(bottom = 4.dp),
+            contentAlignment = Alignment.BottomCenter
+        ) {
+            LText(
+                text = stringResource(R.string.settings),
+                color = MaterialTheme.colors.onPrimary,
+                style = MaterialTheme.typography.title2
+            )
+        }
+    }
+
+    @Composable
+    private fun ColumnScope.Content() {
+        val scrollState = rememberScrollState()
+        val hrmState by PreferencesEx.hrmFeatureConfigStateLd.observeAsState()
+
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1.0f, fill = true)
+                .verticalScroll(scrollState),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            // version info
+            val versionName = packageManager.getPackageInfo(packageName, 0).versionName
+            LText(
+                text = getString(R.string.version_name_X, versionName),
+                modifier = Modifier
+                    .padding(all = 8.dp)
+            )
+
+            // HRM settings
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(start = 16.dp, end = 16.dp)
+                    .clickable {
+                        onHrmClicked()
+                    },
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                LText(
+                    text = stringResource(id = R.string.settings_hrm),
+                    modifier = Modifier
+                        .weight(1.0f)
+                )
+                val hrmEnabled = hrmState === FeatureConfigEnum.ENABLED
+                Icon(
+                    imageVector = ToggleChipDefaults.switchIcon(checked = hrmEnabled),
+                    contentDescription = "",
+                    modifier = Modifier.padding(start = 8.dp),
+                    tint = if (hrmEnabled) MaterialTheme.colors.primary else Color.Gray
+                )
+            }
+            if (PreferencesEx.isDebug) {
+                LText(text = "State: $hrmState")
+            }
+        }
+    }
+
+    private fun onHrmClicked() {
+        when (PreferencesEx.hrmFeatureConfigState) {
+            FeatureConfigEnum.ENABLED -> {
+                PreferencesEx.hrmFeatureConfigState = FeatureConfigEnum.DISABLED
+                if (TrackRecordingService.isRunning()) {
+                    val intent = Intent(this, TrackRecordingService::class.java)
+                    intent.action = TrackRecordingService.ACTION_STOP_FOREGROUND_SERVICE
+                    startService(intent)
                 }
-        )
-    }
-
-    private val isHrmEnabled: Boolean
-        get() {
-            val hrmConfig = PreferencesEx.getHrmFeatureConfig()
-            return hrmConfig === FeatureConfigEnum.ENABLED
-        }
-
-    @Suppress("UNUSED_PARAMETER")
-    fun onHrmClicked(view: View) {
-        if (isHrmEnabled) {
-            PreferencesEx.persistHrmFeatureConfig(FeatureConfigEnum.DISABLED)
-            if (TrackRecordingService.isRunning()) {
-                val intent = Intent(this, TrackRecordingService::class.java)
-                intent.action = TrackRecordingService.ACTION_STOP_FOREGROUND_SERVICE
-                startService(intent)
             }
-        } else {
-            if (RecordingSensorManager.checkAndRequestBodySensorPermission(this)) {
-                enableHrm()
+            else -> {
+                if (BodySensorsManager.checkAndRequestBodySensorsPermission(this)) {
+                    // enable because we have a valid permission
+                    enableHrm()
+                }
             }
         }
-        updateSwitchState()
     }
 
     private fun enableHrm() {
         runOnUiThread {
-            val hrmConfig = PreferencesEx.getHrmFeatureConfig()
-            if (hrmConfig === FeatureConfigEnum.NO_PERMISSION) {
-                if (RecordingSensorManager.checkBodySensorPermission(this)) {
-                    PreferencesEx.persistHrmFeatureConfig(FeatureConfigEnum.NOT_AVAILABLE)
+            val hrmState = PreferencesEx.hrmFeatureConfigState
+            Logger.logD(TAG, "enableHrm(), state: $hrmState")
+
+            // check permission
+            if (hrmState === FeatureConfigEnum.NO_PERMISSION) {
+                if (BodySensorsManager.checkBodySensorsPermission(this)) {
+                    // change just to "not available" to process in next step
+                    PreferencesEx.hrmFeatureConfigState = FeatureConfigEnum.NOT_AVAILABLE
                 } else {
                     Toast.makeText(
-                            this,
-                            getString(R.string.err_no_hrm_permission),
-                            Toast.LENGTH_LONG
+                        this,
+                        getString(R.string.err_no_hrm_permission),
+                        Toast.LENGTH_LONG
                     ).show()
                 }
             }
-            if (hrmConfig === FeatureConfigEnum.NOT_AVAILABLE) {
-                val tmpHrmConfig = RecordingSensorManager.recheckSensorAvailability(this)
+
+            // check availability
+            if (hrmState === FeatureConfigEnum.NOT_AVAILABLE) {
+                val tmpHrmConfig = BodySensorsManager.recheckSensorAvailability(this)
                 // HRM could have been enabled as side effect of calling recheckSensorAvailability()
                 if (tmpHrmConfig === FeatureConfigEnum.NOT_AVAILABLE) {
                     // sensor still not available even after recheck
                     Toast.makeText(
-                            this,
-                            getString(R.string.err_hrm_sensor_not_available),
-                            Toast.LENGTH_LONG
+                        this,
+                        getString(R.string.err_hrm_sensor_not_available),
+                        Toast.LENGTH_LONG
                     ).show()
                 }
             }
-            if (hrmConfig === FeatureConfigEnum.DISABLED) {
+            if (hrmState === FeatureConfigEnum.DISABLED) {
                 // sensor should be accessible and is only disabled, simply enable it
-                PreferencesEx.persistHrmFeatureConfig(FeatureConfigEnum.ENABLED)
+                PreferencesEx.hrmFeatureConfigState = FeatureConfigEnum.ENABLED
             }
         }
     }
 
     override fun onRequestPermissionsResult(
-            requestCode: Int,
-            permissions: Array<String>,
-            grantResults: IntArray) {
+        requestCode: Int,
+        permissions: Array<String>,
+        grantResults: IntArray
+    ) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        RecordingSensorManager.handlePermissionResult(this, requestCode, permissions, grantResults)
+        BodySensorsManager.handlePermissionResult(this, requestCode)
         enableHrm()
-        updateSwitchState()
+    }
+
+    @Preview(
+        widthDp = WEAR_PREVIEW_DEVICE_WIDTH_DP,
+        heightDp = WEAR_PREVIEW_DEVICE_HEIGHT_DP,
+        apiLevel = WEAR_PREVIEW_API_LEVEL,
+        uiMode = WEAR_PREVIEW_UI_MODE,
+        backgroundColor = WEAR_PREVIEW_BACKGROUND_COLOR_BLACK,
+        showBackground = WEAR_PREVIEW_SHOW_BACKGROUND
+    )
+    @Composable
+    fun WearAppPreview() {
+        MainContent()
+    }
+
+    companion object {
+
+        // tag for logger
+        private const val TAG = "MainSettingsActivity"
     }
 }