+++ wear/src/main/java/com/asamm/locus/addon/wear/gui/compose/Components.kt
@@ -1,0 +1,24 @@
+package com.asamm.locus.addon.wear.gui.compose
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.wear.compose.material.LocalTextStyle
+import androidx.wear.compose.material.MaterialTheme
+import androidx.wear.compose.material.Text
+
+@Composable
+fun LText(
+    text: String,
+    modifier: Modifier = Modifier,
+    color: Color = MaterialTheme.colors.onBackground,
+    style: TextStyle = LocalTextStyle.current
+) {
+    Text(
+        text = text,
+        modifier = modifier,
+        color = color,
+        style = style
+    )
+}