+++ wear/src/main/java/com/asamm/locus/addon/wear/gui/compose/theme/Color.kt
@@ -1,0 +1,17 @@
+package com.asamm.locus.addon.wear.gui.compose.theme
+
+import androidx.compose.ui.graphics.Color
+import androidx.wear.compose.material.Colors
+
+val AppColorPalette: Colors = Colors(
+        primary = Color(0xFF1976d2),
+        primaryVariant = Color(0xFF004ba0),
+        secondary = Color(0xFFd84315),
+        secondaryVariant = Color(0xFF9f0000),
+        error = Color(0xFFb00020),
+        background = Color.White,
+        onPrimary = Color.White,
+        onSecondary = Color.White,
+        onError = Color.Black,
+        onBackground = Color.Black
+)