+++ wear/src/main/java/com/asamm/locus/addon/wear/gui/compose/theme/Theme.kt
@@ -1,0 +1,17 @@
+package com.asamm.locus.addon.wear.gui.compose.theme
+
+import androidx.compose.runtime.Composable
+import androidx.wear.compose.material.MaterialTheme
+
+@Composable
+fun AppTheme(
+    content: @Composable () -> Unit
+) {
+    MaterialTheme(
+        colors = AppColorPalette,
+        typography = AppTypography,
+        // For shapes, we generally recommend using the default Material Wear shapes which are
+        // optimized for round and non-round devices.
+        content = content
+    )
+}