+++ wear/src/main/java/com/asamm/locus/addon/wear/gui/compose/theme/Type.kt
@@ -1,0 +1,24 @@
+package com.asamm.locus.addon.wear.gui.compose.theme
+
+import androidx.wear.compose.material.Typography
+
+// Set of Material typography styles to start with
+val AppTypography = Typography(
+        /* Other default text styles to override
+        body1 = TextStyle(
+            fontFamily = FontFamily.Default,
+            fontWeight = FontWeight.Normal,
+            fontSize = 16.sp
+        ),
+        button = TextStyle(
+            fontFamily = FontFamily.Default,
+            fontWeight = FontWeight.W500,
+            fontSize = 14.sp
+        ),
+        caption = TextStyle(
+            fontFamily = FontFamily.Default,
+            fontWeight = FontWeight.Normal,
+            fontSize = 12.sp
+        )
+        */
+)